refactor: 당첨번호를 컴마 기준으로 숫자 배열로 만드는 함수 App.js에서 따로 파일 만들어 이동
