+++ src/App.js
@@ -4,13 +4,18 @@ import {
   getBonusNumberInput,
 } from './Views/inputViews.js';
 import { validatePurchasePrice } from './Models/purchasePriceValidator.js';
+import { countPurchaseAmount } from './Models/purchasePriceUtils.js';
+import {
+  printCountPurchaseAmount,
+  printLottoList,
+  printRateOfReturn,
+} from './Views/outputViews.js';
+import { issueLottoList } from './Models/issueLottoList.js';
 import { validateWinningNumbers } from './Models/winningNumbersValidator.js';
 import { validateBonusNumber } from './Models/bonusNumberValidator.js';
-import { printCountPurchaseAmount, printLottoList, printRateOfReturn } from './Views/outputView.js';
-import { countPurchaseAmount } from './Models/purchasePriceUtils.js';
-import { issueLottoList } from './Models/issueLottoList.js';
 import { produceStatistics } from './Models/winningStatistics.js';
 import { produceRateOfReturn } from './Models/rateOfReturn.js';
+import { trimWinningNumbers } from './Models/winningNumberUtils.js';
 
 class App {
   async run() {
@@ -24,29 +29,14 @@ class App {
     printLottoList(purchaseCount, lottoList);
 
     const winningNumbers = await getWinningNumbersInput();
-
     validateWinningNumbers(winningNumbers);
-    // console.log('winningNumbers: ', winningNumbers);
 
-    // 여기 컴마 뺀 winningNumbers
-    const trimWinningNumbers = (winningNumbers) => {
-      // console.log('winningNumbers: ', winningNumbers);
-      const arr = [...winningNumbers];
-      // console.log('arr111: ', arr);
-      const filtered = arr.filter((element) => element !== ',');
-      // console.log('filtered: ', filtered);
-      const numberedFiltered = filtered.map(Number);
-      // console.log('numberedFiltered: ', numberedFiltered);
-
-      return numberedFiltered;
-    };
-
-    const trimmedWinningNum = trimWinningNumbers(winningNumbers);
+    const trimWinningNum = trimWinningNumbers(winningNumbers);
 
     const bonusNumber = await getBonusNumberInput();
     validateBonusNumber(bonusNumber);
 
-    produceStatistics(trimmedWinningNum, bonusNumber, lottoList);
+    produceStatistics(trimWinningNum, bonusNumber, lottoList);
 
     const rateOfReturn = produceRateOfReturn(purchasePrice);
     printRateOfReturn(rateOfReturn);
