+++ src/Models/winningNumberUtils.js
@@ -1,0 +1,6 @@
+const trimWinningNumbers = (winningNum) => {
+  const trimWinningNum = winningNum.split(',').map((x) => x.trim());
+  return trimWinningNum;
+};
+
+export { trimWinningNumbers };
